refactor: use a file structure that is easier for reusing modules
