+++ pyproject.toml
@@ -17,7 +17,3 @@ biomodals = "biomodals.cli:app"
 [build-system]
 requires = ["uv_build>=0.9,<0.10.0"]
 build-backend = "uv_build"
-
-[tool.uv.build-backend]
-module-name = "biomodals"
-module-root = ""
