+++ src/biomodals/app/bioinfo/gromacs/gmx_mdp/cg.mdp
