+++ src/biomodals/app/bioinfo/paddleocr_app.py
