+++ src/biomodals/service/__init__.py
@@ -1,0 +1,4 @@
+"""Biomodals apps wrapped into web API endpoints.
+
+<https://modal.com/docs/guide/webhooks>
+"""
