+++ src/biomodals/workflow/__init__.py
@@ -1,0 +1,1 @@
+"""Multi-app computational pipelines."""
